% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fermicatsR.R
\docType{data}
\name{pulsars}
\alias{pulsars}
\title{pulsars (Public List of LAT-Detected Gamma-Ray Pulsars)}
\format{A data frame with 7 variables on 161 gamma-ray pulsars:}
\source{
\url{https://confluence.slac.stanford.edu/display/GLAMCOG/Public+List+of+LAT-Detected+Gamma-Ray+Pulsars}
}
\usage{
pulsars
}
\description{
Fermi Large Area Telescope List of Detected Pulsars
https://confluence.slac.stanford.edu/display/GLAMCOG/Public+List+of+LAT-Detected+Gamma-Ray+Pulsars
}
\details{
Last Updated: 2014-11-06
}
\section{Fields}{

\describe{
\item{PSR}{Pulsar name, PSR JHHMM+DDMM, constructed using the RA and Dec}
\item{RAJ_deg}{Right Ascension, J2000, degrees}
\item{DECJ_deg}{Declination, J2000, degrees}
\item{P_ms}{Period, milliseconds}
\item{Edot}{Spin-down luminosity, erg/s}
\item{Codes}{b=Pulsar is in a binary system,
e=Pulsar was detected in gamma rays by EGRET/COMPTEL,
g=Discovered in LAT gamma-ray data,
m=Millisecond pulsar,
p=Pulsar was discovered by the PSC,
r=Discovered in the radio and/or Gamma-ray pulsations detected using the radio ephemeris,
u=Discovered using a Fermi-LAT seed position,
x=Discovered in the x-ray and/or Gamma-ray pulsations detected using the X-ray ephemeris.}
\item{Refs}{References (see web page for details)}
}
}
\keyword{datasets}

